refactor: demoLogic 삭제 및 단순화
