--- server/api/src/main/java/vook/server/api/domain/demo/logic/DemoLogic.java
@@ -1,16 +1,0 @@
-package vook.server.api.domain.demo.logic;
-
-import lombok.RequiredArgsConstructor;
-import vook.server.api.domain.demo.service.DemoTermSearchService;
-import vook.server.api.globalcommon.annotation.DomainLogic;
-
-@DomainLogic
-@RequiredArgsConstructor
-public class DemoLogic {
-
-    private final DemoTermSearchService searchService;
-
-    public DemoTermSearchResult searchTerm(DemoTermSearchCommand params) {
-        return searchService.search(params);
-    }
-}
