+++ server/api/src/main/java/vook/server/api/domain/demo/service/DemoTermSearchService.java
@@ -1,8 +1,5 @@
 package vook.server.api.domain.demo.service;
 
-import vook.server.api.domain.demo.logic.DemoTermSearchCommand;
-import vook.server.api.domain.demo.logic.DemoTermSearchResult;
-
 public interface DemoTermSearchService {
     DemoTermSearchResult search(DemoTermSearchCommand params);
 }
