+++ server/api/src/main/java/vook/server/api/infra/search/demo/MeilisearchDemoTermSearchService.java
@@ -9,10 +9,10 @@ import com.meilisearch.sdk.model.TaskInfo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.stereotype.Service;
-import vook.server.api.domain.demo.logic.DemoTermSearchCommand;
-import vook.server.api.domain.demo.logic.DemoTermSearchResult;
 import vook.server.api.domain.demo.model.DemoTerm;
 import vook.server.api.domain.demo.model.DemoTermSynonym;
+import vook.server.api.domain.demo.service.DemoTermSearchCommand;
+import vook.server.api.domain.demo.service.DemoTermSearchResult;
 import vook.server.api.domain.demo.service.DemoTermSearchService;
 import vook.server.api.infra.search.common.MeilisearchProperties;
 import vook.server.api.infra.search.common.MeilisearchService;
