+++ server/api/src/test/java/vook/server/api/infra/search/demo/MeilisearchDemoTermSearchServiceTest.java
@@ -1,4 +1,4 @@
-package vook.server.api.domain.demo.logic;
+package vook.server.api.infra.search.demo;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -10,7 +10,8 @@ import vook.server.api.devhelper.app.InitService;
 import vook.server.api.devhelper.app.TestTermsLoader;
 import vook.server.api.domain.demo.model.DemoTerm;
 import vook.server.api.domain.demo.model.DemoTermRepository;
-import vook.server.api.infra.search.demo.MeilisearchDemoTermSearchService;
+import vook.server.api.domain.demo.service.DemoTermSearchCommand;
+import vook.server.api.domain.demo.service.DemoTermSearchResult;
 import vook.server.api.testhelper.IntegrationTestBase;
 
 import java.util.List;
@@ -19,16 +20,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @Transactional
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class DemoLogicTest extends IntegrationTestBase {
+class MeilisearchDemoTermSearchServiceTest extends IntegrationTestBase {
+
     @Autowired
-    private DemoLogic demoLogic;
+    MeilisearchDemoTermSearchService service;
 
     @Autowired
     private TestTermsLoader testTermsLoader;
     @Autowired
     private DemoTermRepository demoTermRepository;
-    @Autowired
-    private MeilisearchDemoTermSearchService demoTermSearchService;
 
     @BeforeAll
     void beforeAll() {
@@ -37,17 +37,17 @@ class DemoLogicTest extends IntegrationTestBase {
                 InitService::convertToDemoTerm
         );
         demoTermRepository.saveAll(terms);
-        demoTermSearchService.init();
-        demoTermSearchService.addTerms(terms);
+        service.init();
+        service.addTerms(terms);
     }
 
     @AfterAll
     void afterAll() {
-        demoTermSearchService.clearAll();
+        service.clearAll();
     }
 
     @Test
-    void searchTerm() {
+    void search() {
         DemoTermSearchCommand params = DemoTermSearchCommand.builder()
                 .query("하이브리드앱")
                 .withFormat(false)
@@ -55,7 +55,7 @@ class DemoLogicTest extends IntegrationTestBase {
                 .highlightPostTag("</em>")
                 .build();
 
-        DemoTermSearchResult result = demoLogic.searchTerm(params);
+        DemoTermSearchResult result = service.search(params);
 
         assertThat(result).isNotNull();
         assertThat(result.query()).isEqualTo("하이브리드앱");
